feat: 設定ページのスタイル統一とAPI Keysをシークレットに移動
- 接続ページからAPI Keysセクションを削除しGitHub連携のみに
- シークレットページにAPI Keys管理を移動
- 全設定ページのフォントサイズ・間隔をメインページに統一
- 自動研究ナビをクリック不可に変更
- プラン・領収書ページのテキストを日本語化
