+++ frontend/src/components/pages/receipts.tsx
@@ -1,10 +1,14 @@
 export function ReceiptsPage() {
   return (
-    <div className="flex-1 p-8 overflow-y-auto">
-      <h1 className="text-2xl font-bold text-foreground">領収書 / 請求書</h1>
-      <p className="mt-2 text-sm text-muted-foreground">過去の領収書や請求書を確認できます。</p>
-      <div className="mt-8 rounded-lg border border-border p-6 text-center text-sm text-muted-foreground">
-        領収書・請求書はまだありません。
+    <div className="flex-1 overflow-y-auto">
+      <div className="max-w-2xl mx-auto px-8 py-8">
+        <h1 className="text-heading-2 font-heading-2 text-default-font">領収書 / 請求書</h1>
+        <p className="text-caption font-caption text-subtext-color mt-1">
+          過去の領収書や請求書を確認できます。
+        </p>
+        <div className="mt-6 rounded-lg border border-border bg-card p-5 text-center text-body font-body text-subtext-color">
+          領収書・請求書はまだありません。
+        </div>
       </div>
     </div>
   );
